refactor(ui): extract render functions into dedicated ui module
- layout, components, history, collections, popups
