+++ src/ui/layout.rs
@@ -1,0 +1,16 @@
+use ratatui::layout::Rect;
+
+pub fn center_rect(width: u16, height: u16, area: Rect) -> Rect {
+    let vertical = Rect::new(
+        area.x,
+        area.y + (area.height.saturating_sub(height)) / 2,
+        area.width.min(width),
+        height.min(area.height),
+    );
+    Rect::new(
+        area.x + (area.width.saturating_sub(width)) / 2,
+        vertical.y,
+        vertical.width,
+        vertical.height,
+    )
+}
